+++ tests/tests_unit/tests_common/test_utils.py
@@ -1,9 +1,9 @@
 import pytest
 from datetime import datetime
-from th2_data_services.data_source.lwdp.utils import _check_datetime
+from th2_data_services.data_source.lwdp.utils import _check_timestamp
 
 
 def test_datetime_invalid_type():
     with pytest.raises(Exception) as err:
-        _check_datetime(datetime.now().timestamp())
-    assert "Provided timestamp should be `datetime` object" in str(err)
+        _check_timestamp(datetime.now().timestamp())
+    assert "Provided timestamp should be `datetime`, `str` or `int` object in UTC time" in str(err)